Chore: replace block.Upload() implementation with Mimir one
