+++ pkg/ingester/shipper.go
@@ -23,7 +23,6 @@ import (
 	"github.com/prometheus/prometheus/tsdb/fileutil"
 	"github.com/thanos-io/objstore"
 
-	"github.com/grafana/mimir/pkg/storage/tsdb"
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
 	"github.com/grafana/mimir/pkg/storage/tsdb/metadata"
 )
@@ -199,7 +198,7 @@ func (s *Shipper) upload(ctx context.Context, meta *metadata.Meta) error {
 	meta.Thanos.SegmentFiles = block.GetSegmentFiles(blockDir)
 
 	// Upload block with custom metadata.
-	return tsdb.UploadBlock(ctx, s.logger, s.bucket, blockDir, meta)
+	return block.Upload(ctx, s.logger, s.bucket, blockDir, meta)
 }
 
 // blockMetasFromOldest returns the block meta of each block found in dir
