+++ pkg/storegateway/indexheader/reader_pool_test.go
@@ -58,7 +58,7 @@ func TestReaderPool_NewBinaryReader(t *testing.T) {
 		labels.FromStrings("a", "2"),
 	}, 100, 0, 1000, labels.FromStrings("ext1", "1"), 124, metadata.NoneFunc)
 	require.NoError(t, err)
-	require.NoError(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), metadata.NoneFunc))
+	require.NoError(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), nil))
 
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
@@ -96,7 +96,7 @@ func TestReaderPool_ShouldCloseIdleLazyReaders(t *testing.T) {
 		labels.FromStrings("a", "2"),
 	}, 100, 0, 1000, labels.FromStrings("ext1", "1"), 124, metadata.NoneFunc)
 	require.NoError(t, err)
-	require.NoError(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), metadata.NoneFunc))
+	require.NoError(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), nil))
 
 	metrics := NewReaderPoolMetrics(nil)
 	pool := NewReaderPool(log.NewNopLogger(), true, idleTimeout, metrics)
